Refactor Dtos to dto packages
