+++ backend/src/main/java/de/neuefische/greensitter/api/dtos/SearchResultData.java
@@ -1,4 +1,4 @@
-package de.neuefische.greensitter.api;
+package de.neuefische.greensitter.api.dtos;
 
 
 import lombok.AllArgsConstructor;
@@ -8,11 +8,11 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class PlantData {
+public class SearchResultData {
 
     private String id;
-    private String scientific_name;
     private String common_name;
+    private String scientific_name;
     private String genus;
     private String family_common_name;
     private String image_url;
