+++ backend/src/main/java/de/neuefische/greensitter/api/dtos/TrefleApiQueryDto.java
@@ -1,17 +1,16 @@
-package de.neuefische.greensitter.api;
+package de.neuefische.greensitter.api.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class TrefleApiQueryResults {
+public class TrefleApiQueryDto {
 
 
-        private PlantData[] data;
+        private SearchResultData[] data;
 
 
     }
