+++ backend/src/main/java/de/neuefische/greensitter/model/dtos/LoginData.java
@@ -1,4 +1,4 @@
-package de.neuefische.greensitter.model;
+package de.neuefische.greensitter.model.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
